refactor: Moving article warehouse module the new architecture
