+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "store-vue",
+  "name": "warehouse-vue",
   "version": "0.0.0",
   "scripts": {
     "dev": "vite",
