+++ src/modules/Warehouse/views/ArticleWarehouse/components/ArticleSearchTable/composables/useTableGrid.ts
@@ -1,7 +1,7 @@
 import { onMounted, watch, inject } from "vue"
 import { useStoreWarehouse } from "@/modules/Warehouse/stores/"
-import { getArticlesSearchByCategories } from "../../../services/Article"
-import WarehouseService from  "../../../services/Warehouse"
+import { getArticlesSearchByCategories } from "@/modules/Warehouse/services/Article"
+import WarehouseService from  "@/modules/Warehouse/services/Warehouse"
 import type {
   TableGrid,
   Data as TableGridData,
