+++ src/modules/Warehouse/views/Warehouse/CreateOrEdit/composables/useCreateOrEdit.ts
@@ -1,9 +1,9 @@
 import { onMounted, reactive, ref } from 'vue'
 import { useRouter } from 'vue-router'
 import useHttp from "@/core/composables/useHttp"; //foreign_table_name
-import WarehouseService from "../../services/Warehouse"
-import ArticleService from "../../services/Article"
-import type { Warehouse } from "../../types/Warehouse"
+import WarehouseService from "@/modules/Warehouse/services/Warehouse"
+import ArticleService from "@/modules/Warehouse/services/Article"
+import type { Warehouse } from "@/modules/Warehouse/types/Warehouse"
 
 export default (propsId?: string) => {
   const router = useRouter();
